server/requirements.txt
anyio==4.11.0
babel==2.17.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
courlan==1.3.2
dateparser==1.2.2
dotenv==0.9.9
filelock==3.20.0
Flask==3.1.2
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.11
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tld==0.13.1
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.57.3
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
